Move dummy server creation to testing utilities.
