+++ src/lib/testUtils.ts
@@ -1,0 +1,30 @@
+import http from "http";
+
+const host = "127.0.0.1";
+const port = 8080;
+const url = `http://${host}:${port}/`;
+
+const createServer = (): Promise<http.Server> =>
+	new Promise<http.Server>((resolve, reject) => {
+		const server = http.createServer().listen(port, host);
+
+		server.on("listening", () => resolve(server));
+		server.on("error", reject);
+	});
+
+const closeServer = (server: http.Server): Promise<void> =>
+	new Promise<void>((resolve, reject) => {
+		if (server.listening) {
+			server.close((error) => {
+				if (error) {
+					reject(error);
+				} else {
+					resolve();
+				}
+			});
+		} else {
+			resolve();
+		}
+	});
+
+export {host, port, url, createServer, closeServer};
